modifico la forma de borrar partidas incompletas

+++ apps/trivia_game/views.py
@@ -89,7 +89,8 @@ class ResultsGameView(View):
         # Elimina todas aquellas partidas que no llegaron a tu termino para que no
         # las incluya en el ranking de partidas completadas
         incomplete_games = Trivia_game.objects.filter(time__isnull=True)
-        incomplete_games.delete()
+        if incomplete_games:
+            incomplete_games.delete()
     
         # aqui busco un rankin de las 6 mejores partidas, pero por el momento no estoy contemplando 
         # que los usuarios se van a estar repitiendo con el mismo nombre.
